adding codes for updating students and error handling

+++ src/main/java/com/wileyedge/studentwebservice/service/UserService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import com.wileyedge.studentwebservice.dao.IDao;
+import com.wileyedge.studentwebservice.exception.StudentNotFoundException;
 import com.wileyedge.studentwebservice.model.Student;
 
 @Service
@@ -41,16 +42,46 @@ public class UserService implements IService {
 	}
 
 	@Override
-	public void deleteOneStudent(int id) {
-		dao.deleteById(id);		
+	public List<Student> findByName(String name){
+		List<Student> students = dao.findByName(name);
+		if(students == null || students.isEmpty()) {
+			throw new StudentNotFoundException("Student named " + name + " does not exist.");
+		}
+		return students;
 	}
 	
 	@Override
-	public List<Student> findByName(String name){
-		List<Student> students = dao.findByName(name);
+	public void deleteOneStudent(int id) {
+		Optional<Student> existingStudentOptional = dao.findById(id);
+		if(existingStudentOptional == null || existingStudentOptional.isEmpty()) {
+			throw new StudentNotFoundException("Student with ID " + id + " does not exist.");
+		}else {
+			dao.deleteById(id);		
+		}
+	}
+	
+	@Override
+	public Student updateStudent(Student student) {	
+		Student existingStudent = null;
 		
-		return students;
+		// Check if the student exists
+		Optional<Student> existingStudentOptional = dao.findById(student.getStuid());
+		if (!existingStudentOptional.isPresent()) {
+			throw new StudentNotFoundException("Student with ID " + student.getStuid() + " does not exist.");
+		}else {
+			existingStudent = existingStudentOptional.get();
+		}
 		
+		//Update existing student
+		existingStudent.setName(student.getName());
+		existingStudent.setAge(student.getAge());
+		existingStudent.setMob(student.getMob());
+		existingStudent.setAddr(student.getAddr());
+		
+		//Save updated student info to db
+		Student updatedStudent = dao.save(existingStudent);
+		
+		return updatedStudent;
 	}
 
 }
